+++ fon/TimeSoundEditor.cpp
@@ -683,7 +683,7 @@ bool structTimeSoundEditor :: v_clickB (double xbegin, double ybegin) {
 void TimeSoundEditor_init (TimeSoundEditor me, const char32 *title, Function data, Sampled sound, bool ownSound) {
 	my d_ownSound = ownSound;
 	if (sound) {
-		long numberOfChannels = 1;
+		integer numberOfChannels = 1;
 		if (ownSound) {
 			Melder_assert (Thing_isa (sound, classSound));
 			my d_sound.data = Data_copy ((Sound) sound).releaseToAmbiguousOwner();   // deep copy; ownership transferred
